+++ osidb/models/__init__.py
@@ -1,7 +1,7 @@
 from .abstract import CVSS, ComparableTextChoices, Impact
 from .flaw import FlawSource
 from .profile import Profile
-from .ps_constants import SpecialConsiderationPackage, UbiPackage
+from .ps_constants import SpecialConsiderationPackage
 from .ps_contact import PsContact
 from .ps_module import PsModule
 from .ps_product import PsProduct
@@ -31,7 +31,6 @@ __all__ = (
     "Snippet",
     "SpecialConsiderationPackage",
     "Tracker",
-    "UbiPackage",
 )
 
 from .affect import Affect, AffectCVSS